Added Batch Rendering for flat colored Quads

+++ Engine/src/Renderer/Buffer.h
@@ -113,13 +113,13 @@ namespace Alexio
 		virtual void Bind() const = 0;
 		virtual void Unbind() const = 0;
 
-		virtual void SetData(const void* data, uint32_t size) = 0;
+		virtual void SetData(const void* data, size_t size) = 0;
 
 		void SetLayout(const BufferLayout& layout) { mLayout = layout; }
 		const BufferLayout& GetLayout() const { return mLayout; }
 
-		static Ref<VertexBuffer> Create(uint32_t size);
-		static Ref<VertexBuffer> Create(const void* data, uint32_t size);
+		static Ref<VertexBuffer> Create(size_t size);
+		static Ref<VertexBuffer> Create(const void* data, size_t size);
 	protected:
 		BufferLayout mLayout;
 	};
@@ -133,9 +133,9 @@ namespace Alexio
 		virtual void Unbind() = 0;
 
 		inline uint32_t GetCount() const { return mCount; }
-		static Ref<IndexBuffer> Create(uint32_t* indices, uint32_t count);
+		static Ref<IndexBuffer> Create(uint32_t* indices, size_t count);
 	protected:
-		uint32_t mCount;
+		size_t mCount;
 	};
 
 	class VertexArray
