+++ Engine/src/Renderer/OpenGL/OpenGL_Buffer.cpp
@@ -5,14 +5,14 @@
 
 namespace Alexio
 {
-	OpenGL_VertexBuffer::OpenGL_VertexBuffer(uint32_t size)
+	OpenGL_VertexBuffer::OpenGL_VertexBuffer(size_t size)
 	{
 		glCreateBuffers(1, &mID);
 		glBindBuffer(GL_ARRAY_BUFFER, mID);
 		glBufferData(GL_ARRAY_BUFFER, size, nullptr, GL_DYNAMIC_DRAW);
 	}
 
-	OpenGL_VertexBuffer::OpenGL_VertexBuffer(const void* data, uint32_t size)
+	OpenGL_VertexBuffer::OpenGL_VertexBuffer(const void* data, size_t size)
 	{
 		glCreateBuffers(1, &mID);
 		glBindBuffer(GL_ARRAY_BUFFER, mID);
@@ -34,7 +34,7 @@ namespace Alexio
 		glBindBuffer(GL_ARRAY_BUFFER, 0);
 	}
 
-	void OpenGL_VertexBuffer::SetData(const void* data, uint32_t size)
+	void OpenGL_VertexBuffer::SetData(const void* data, size_t size)
 	{
 		glBindBuffer(GL_ARRAY_BUFFER, mID);
 		glBufferSubData(GL_ARRAY_BUFFER, 0, size, data);
@@ -44,7 +44,7 @@ namespace Alexio
 	///////////INDEX BUFFER//////////////////////////////
 	/////////////////////////////////////////////////////
 
-	OpenGL_IndexBuffer::OpenGL_IndexBuffer(uint32_t* indices, uint32_t count)
+	OpenGL_IndexBuffer::OpenGL_IndexBuffer(uint32_t* indices, size_t count)
 	{
 		mCount = count;
